requirements.txt
fastapi
uvicorn
requests
google-auth
google-auth-oauthlib
google-api-python-client
google-generativeai
